requirements.txt
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
torch>=2.0.0
pyyaml>=6.0
datasets>=2.14.0
transformers>=4.35.0
accelerate>=0.24.0
pytest>=7.0.0
